requirements.txt
gnosis-py==3.9.3
matplotlib==3.5.1
pre-commit==2.18.1
python-dotenv==0.20.0
